requirements.txt
asgiref==3.2.10
boto3==1.15.1
botocore==1.18.1
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.1
django-crum==0.7.7
django-extensions==3.1.3
django-filter==1.1.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-storages==1.10.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.4
Pillow==7.2.0
psycopg2==2.8.6
pydot==1.4.2
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
simplejson==3.17.5
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
